moved errors header to appropriate folder
