+++ control/EventLoop.cpp
@@ -1,6 +1,6 @@
 #include "control/EventLoop.h"
 
-#include "error.h"
+#include "errors/ErrorConstants.h"
 
 EventLoop :: EventLoop(const GameField& field, const InputList& inputs)
 	: m_field (field)
